telegram bot: init function was moved to separate file
